profile page and dropzone added

+++ database/migrations/2023_02_11_043042_add_property_id_to_users_table.php
@@ -14,8 +14,10 @@ class AddPropertyIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('property_id')->after('id');
-            $table->foreign('property_id')->references('id')->on('properties');
+            if (!Schema::hasColumn('users', 'property_id')) {
+                $table->unsignedBigInteger('property_id')->after('id');
+                $table->foreign('property_id')->references('id')->on('properties')->onDelete('cascade');
+            }
         });
     }
 
@@ -27,7 +29,9 @@ class AddPropertyIdToUsersTable extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('property_id');
+            if (!Schema::hasColumn('users', 'property_id')) {
+                $table->dropColumn('property_id');
+            }
         });
     }
 }
